Phase 4: lift CatalogCard out of ArcadeCatalog
ArcadeCatalog.tsx was 1111 lines and more than a third of them were one
sub-component: the grid card, with its own hover-intent timers, preview
video lifecycle and badge stack. It already took explicit props and
closed over nothing from the surface around it, so it moves whole.

CatalogCard.tsx also carries humanizeMoment and HOVER_INTENT_MS, the two
things only the card used.

The comments travelling with it are compressed to the one-line form the
prose gate wants: the badge-stacking note, the label-span note, the
touch-warning note and the contributor-credit block were paragraphs, and
each says the same thing in a single line now. The JSX ones become `//`
inside their expression container, since a block comment is prose
whatever it is wrapped in.

ArcadeCatalog 1111 -> 678, and its prose baseline 830 -> 258. No
behaviour change; the card renders the same DOM from the same props.
