+++ apps/web/src/surfaces/catalog/ArcadeCatalog.tsx
@@ -3,13 +3,8 @@ import { useTranslation } from 'react-i18next';
 import './catalog-toolbar.css';
 import './catalog.css';
 import { useAuth } from '../../AuthContext.js';
-import {
-  catalogMediaUrl,
-  defaultScreenshotIndex,
-  gamePageHandle,
-  isPlatformAuthor,
-  type CatalogEntry,
-} from '../../catalog.js';
+import { type CatalogEntry } from '../../catalog.js';
+import { CatalogCard } from './CatalogCard.js';
 import {
   CATALOG_CATEGORY_IDS,
   categorizeCatalogEntry,
@@ -38,14 +33,12 @@ import { fetchFeaturedSlugs } from '../../featuredApi.js';
 import { MascotMoment } from '../../Mascot.js';
 import { PixelIcon } from '../../PixelIcon.js';
 import { getRecentPlays } from '../../recentPlays.js';
-import { creatorPath, gamePath } from '../../core/router.js';
 import {
   fetchCatalogSortSignals,
   readCachedCatalogSortPayload,
   type CatalogSortPayload,
 } from '../../recommendationsApi.js';
-import { useInView } from '../../useInView.js';
-import { isCatalogScrolling, watchCatalogScrollIdle, whenCatalogScrollIdle } from './catalogScrollIdle.js';
+import { watchCatalogScrollIdle } from './catalogScrollIdle.js';
 import type { PlayVia } from '../../visitTelemetry.js';
 
 // Rail length before it scrolls, same for every rail.
@@ -66,432 +59,6 @@ type ArcadeCatalogProps = {
   /** Bump after a new submission so Yours pins refresh. */
   creatorGamesRefreshKey?: number;
 };
-
-function humanizeMoment(name: string): string {
-  return name
-    .split('-')
-    .map((word) => `${word.charAt(0).toUpperCase()}${word.slice(1)}`)
-    .join(' ');
-}
-
-/** Hover must dwell this long before we fetch moments / arm video — scroll sweeps skip it. */
-const HOVER_INTENT_MS = 240;
-
-function CatalogCard({
-  entry,
-  isYours = false,
-  onPlayGame,
-  onPlayTogether,
-}: {
-  entry: CatalogEntry;
-  isYours?: boolean;
-  onPlayGame: (game: CatalogEntry) => void;
-  onPlayTogether: (game: CatalogEntry) => void;
-}) {
-  const { t } = useTranslation();
-  const videoRef = useRef<HTMLVideoElement>(null);
-  const hoverTimerRef = useRef<number | null>(null);
-  const hoveringRef = useRef(false);
-  const cancelIdleWaitRef = useRef<(() => void) | null>(null);
-  // Poster when near the fold; video/moments only after deliberate engage (dwell hover,
-  // keyboard focus, or play). Leave-view unloads so scrolled-away cards stay light.
-  const { ref: mediaRef, inView } = useInView<HTMLDivElement>({ rootMargin: '80px 0px', once: false });
-  const screenshots = entry.media?.screenshots ?? [];
-  const [selectedScreenshot, setSelectedScreenshot] = useState(() => defaultScreenshotIndex(screenshots));
-  const [isPreviewPlaying, setIsPreviewPlaying] = useState(false);
-  const [isPreviewPinned, setIsPreviewPinned] = useState(false);
-  const [videoArmed, setVideoArmed] = useState(false);
-  const [extrasOpen, setExtrasOpen] = useState(false);
-  const selected = screenshots[selectedScreenshot] ?? screenshots[0];
-  // 640 rather than the original: the card box is a few hundred CSS pixels, so this is
-  // roughly 2x what is drawn and a fraction of a full screenshot to decode.
-  const posterUrl = selected && inView ? catalogMediaUrl(entry.slug, selected.file, 640) : undefined;
-  const hasVideo = Boolean(entry.media?.video);
-  const hasMoments = screenshots.length > 1;
-  const videoUrl =
-    hasVideo && inView && videoArmed && entry.media?.video ? catalogMediaUrl(entry.slug, entry.media.video) : null;
-  const showMoments = extrasOpen && inView && hasMoments;
-
-  useEffect(() => {
-    if (inView) return;
-    hoveringRef.current = false;
-    if (hoverTimerRef.current != null) {
-      window.clearTimeout(hoverTimerRef.current);
-      hoverTimerRef.current = null;
-    }
-    cancelIdleWaitRef.current?.();
-    cancelIdleWaitRef.current = null;
-    setExtrasOpen(false);
-    setVideoArmed(false);
-    setIsPreviewPlaying(false);
-    setIsPreviewPinned(false);
-  }, [inView]);
-
-  useEffect(() => {
-    if (!videoUrl) return;
-    const video = videoRef.current;
-    if (!video) return;
-    try {
-      void Promise.resolve(video.play()).then(
-        () => setIsPreviewPlaying(true),
-        () => setIsPreviewPlaying(false),
-      );
-    } catch {
-      setIsPreviewPlaying(false);
-    }
-  }, [videoUrl]);
-
-  useEffect(
-    () => () => {
-      hoveringRef.current = false;
-      if (hoverTimerRef.current != null) window.clearTimeout(hoverTimerRef.current);
-      cancelIdleWaitRef.current?.();
-      cancelIdleWaitRef.current = null;
-    },
-    [],
-  );
-
-  function armPreview() {
-    setVideoArmed(true);
-    const video = videoRef.current;
-    if (!video?.src) return;
-    try {
-      void Promise.resolve(video.play()).then(
-        () => setIsPreviewPlaying(true),
-        () => setIsPreviewPlaying(false),
-      );
-    } catch {
-      setIsPreviewPlaying(false);
-    }
-  }
-
-  function stopPreview() {
-    const video = videoRef.current;
-    if (video) {
-      video.pause();
-      video.removeAttribute('src');
-      video.load();
-    }
-    setVideoArmed(false);
-    setIsPreviewPlaying(false);
-  }
-
-  function pausePreview() {
-    const video = videoRef.current;
-    if (video) video.pause();
-    setIsPreviewPlaying(false);
-  }
-
-  function clearHoverIntent() {
-    if (hoverTimerRef.current != null) {
-      window.clearTimeout(hoverTimerRef.current);
-      hoverTimerRef.current = null;
-    }
-    cancelIdleWaitRef.current?.();
-    cancelIdleWaitRef.current = null;
-  }
-
-  /** Open moments (and arm video when present) once the pointer is idle over the card. */
-  function armExtrasFromHover() {
-    if (!hoveringRef.current) return;
-    if (isCatalogScrolling()) {
-      // Inertial scroll can leave the cursor parked on the same card — wait for idle
-      // instead of discarding the intent until the user re-enters.
-      cancelIdleWaitRef.current?.();
-      cancelIdleWaitRef.current = whenCatalogScrollIdle(() => {
-        cancelIdleWaitRef.current = null;
-        armExtrasFromHover();
-      });
-      return;
-    }
-    setExtrasOpen(true);
-    if (hasVideo) armPreview();
-  }
-
-  function scheduleHoverIntent() {
-    clearHoverIntent();
-    hoveringRef.current = true;
-    hoverTimerRef.current = window.setTimeout(() => {
-      hoverTimerRef.current = null;
-      armExtrasFromHover();
-    }, HOVER_INTENT_MS);
-  }
-
-  function endHoverIntent() {
-    hoveringRef.current = false;
-    clearHoverIntent();
-    if (isPreviewPinned) return;
-    setExtrasOpen(false);
-    stopPreview();
-  }
-
-  function togglePreview() {
-    hoveringRef.current = false;
-    clearHoverIntent();
-    if (isPreviewPlaying) {
-      setIsPreviewPinned(false);
-      pausePreview();
-    } else {
-      setExtrasOpen(true);
-      setIsPreviewPinned(true);
-      armPreview();
-    }
-  }
-
-  /** Video-less cards have no play toggle — pin/unpin the moment strip explicitly. */
-  function toggleMoments() {
-    hoveringRef.current = false;
-    clearHoverIntent();
-    if (extrasOpen) {
-      setIsPreviewPinned(false);
-      setExtrasOpen(false);
-    } else {
-      setExtrasOpen(true);
-      setIsPreviewPinned(true);
-    }
-  }
-
-  function selectScreenshot(index: number) {
-    setSelectedScreenshot(index);
-    setIsPreviewPinned(false);
-    stopPreview();
-  }
-
-  return (
-    <article className="catalog-card">
-      <div
-        ref={mediaRef}
-        className="catalog-media"
-        tabIndex={hasVideo || hasMoments ? 0 : undefined}
-        onPointerEnter={(event) => {
-          if (event.pointerType === 'mouse') scheduleHoverIntent();
-        }}
-        onPointerLeave={(event) => {
-          if (event.pointerType === 'mouse') endHoverIntent();
-        }}
-        onFocus={(event) => {
-          if (event.target !== event.currentTarget) return;
-          setExtrasOpen(true);
-          if (hasVideo) armPreview();
-        }}
-        onBlur={(event) => {
-          if (event.currentTarget.contains(event.relatedTarget)) return;
-          hoveringRef.current = false;
-          clearHoverIntent();
-          setIsPreviewPinned(false);
-          setExtrasOpen(false);
-          stopPreview();
-        }}
-      >
-        <a
-          className="catalog-card-hit-area"
-          href={gamePath(gamePageHandle(entry), entry.slug)}
-          aria-label={`${entry.title} — ${t('catalog.openGame')}`}
-        />
-        {videoUrl ? (
-          <video
-            ref={videoRef}
-            className="catalog-preview"
-            src={videoUrl}
-            poster={posterUrl}
-            muted
-            loop
-            playsInline
-            preload="auto"
-            aria-label={t('catalog.previewVideo', { title: entry.title })}
-            onPlay={() => setIsPreviewPlaying(true)}
-            onPause={() => setIsPreviewPlaying(false)}
-          />
-        ) : posterUrl ? (
-          <img
-            className="catalog-preview"
-            src={posterUrl}
-            alt={t('catalog.previewImage', { title: entry.title })}
-            loading="lazy"
-            decoding="async"
-          />
-        ) : (
-          <div className="catalog-preview-fallback" aria-hidden="true">
-            <span>{entry.title.charAt(0)}</span>
-            <small>{entry.genre}</small>
-          </div>
-        )}
-
-        {/* Both badges anchor the same corner, so they stack in a column instead of
-            each claiming top:10px/left:10px independently — the previous version had
-            the AI disclosure drawn directly on top of the preview toggle on every
-            card with a video. Flex + gap rather than a fixed pixel offset because the
-            toggle's Polish label ("Wstrzymaj podgląd") runs longer than the English
-            one and can wrap; a magic number here would work in one locale and not
-            the other. */}
-        <div className="catalog-badges-top-left">
-          {/* The label is in its own span because a finger gets the icon alone: at
-              44px the labelled pill is 127px wide (more in Polish) and ran into the
-              moment thumbnails on the other side of the card. aria-label carries the
-              wording that the CSS hides, so the control never goes unnamed. */}
-          {hasVideo && (
-            <button
-              type="button"
-              className="preview-toggle preview-toggle--video"
-              aria-pressed={isPreviewPlaying}
-              aria-label={isPreviewPlaying ? t('catalog.pausePreview') : t('catalog.watchPreview')}
-              disabled={!inView}
-              onClick={togglePreview}
-            >
-              <PixelIcon name={isPreviewPlaying ? 'pause' : 'play'} size={11} />
-              <span className="btn-label">
-                {isPreviewPlaying ? t('catalog.pausePreview') : t('catalog.watchPreview')}
-              </span>
-            </button>
-          )}
-
-          {/* Cards without a trailer still need a deliberate, accessible way to open
-              the moment strip — hover dwell is mouse-only and media focus is easy to
-              miss on touch. Reuse the preview-toggle chrome so the badge column stays
-              one shape. */}
-          {!hasVideo && hasMoments && (
-            <button
-              type="button"
-              className="preview-toggle"
-              aria-pressed={extrasOpen}
-              aria-label={extrasOpen ? t('catalog.hideMoments') : t('catalog.showMoments')}
-              disabled={!inView}
-              onClick={toggleMoments}
-            >
-              <PixelIcon name="image" size={11} />
-              <span className="btn-label">{extrasOpen ? t('catalog.hideMoments') : t('catalog.showMoments')}</span>
-            </button>
-          )}
-
-          {/* The one thing a phone visitor needs from the derived `touch` class: a game
-              its own source says cannot be driven by a finger. Only 'none' earns a
-              badge — 'gamekit'/'native' are the norm for nearly every game, so labelling
-              those would put a pill on almost every card to say "works normally", while
-              'controllers' already announces itself through the party badge below.
-              Hidden on a mouse by CSS, where "keyboard only" is not news. */}
-          {entry.touch === 'none' && (
-            <span className="touch-warning-pill" title={t('catalog.keyboardOnlyTooltip')}>
-              <PixelIcon name="gamepad" size={10} /> {t('catalog.keyboardOnly')}
-            </span>
-          )}
-
-          {isYours ? (
-            <span className="yours-pill" title={t('catalog.yoursBadge')}>
-              <PixelIcon name="user" size={10} /> {t('catalog.yoursBadge')}
-            </span>
-          ) : null}
-        </div>
-
-        <span className="genre-pill">{entry.genre}</span>
-
-        {showMoments && (
-          <div className="catalog-moments" aria-label={t('catalog.gameMoments', { title: entry.title })}>
-            {screenshots.slice(0, 4).map((screenshot, index) => (
-              <button
-                key={screenshot.file}
-                type="button"
-                className={index === selectedScreenshot ? 'catalog-moment is-selected' : 'catalog-moment'}
-                aria-label={t('catalog.viewMoment', { moment: humanizeMoment(screenshot.name), title: entry.title })}
-                aria-pressed={index === selectedScreenshot}
-                onClick={() => selectScreenshot(index)}
-              >
-                <img src={catalogMediaUrl(entry.slug, screenshot.file, 96)} alt="" loading="lazy" decoding="async" />
-              </button>
-            ))}
-          </div>
-        )}
-
-        {/* Title, one-line control hint and the Play CTA sit on top of the preview over
-            a bottom scrim, so the card is just the media box — no separate content
-            row underneath. Saves ~90px of height per card without dropping any info. */}
-        <div className="catalog-overlay">
-          <div className="card-copy">
-            <h3 className="card-title">
-              {entry.title}
-              {entry.multiplayer && (
-                <span className="card-party-badge">
-                  <PixelIcon name="phone" size={12} /> {t('party.playersBadge', { max: entry.multiplayer.maxPlayers })}
-                </span>
-              )}
-              {/* Says what the game will do, not what this visitor will get: a signed-out
-                player sees the badge and no save, which is the honest ordering — the
-                promise belongs to the game, and signing in is what claims it. */}
-              {entry.saves === 'player' && (
-                <span className="card-saves-badge">
-                  <PixelIcon name="clock" size={12} /> {t('catalog.savesBadge')}
-                </span>
-              )}
-              {/* The one badge that is about other people rather than about the game.
-                Worth its own colour: "somebody else has been here" is a different kind
-                of reason to click than "this remembers you". */}
-              {entry.world === 'shared' && (
-                <span className="card-world-badge">
-                  <PixelIcon name="star" size={12} /> {t('catalog.worldBadge')}
-                </span>
-              )}
-              {/* Advisory, like saves: says the game answers tilt where the device offers
-                it, while the keyboard stays the whole game everywhere else. Reuses the
-                party pill style — it is the same "how you can drive this" family. */}
-              {entry.sensing === 'tilt' && (
-                <span className="card-party-badge" title={t('catalog.tiltTooltip')}>
-                  <PixelIcon name="phone" size={12} /> {t('catalog.tiltBadge')}
-                </span>
-              )}
-              {entry.sensing === 'backdrop' && (
-                <span className="card-party-badge" title={t('catalog.cameraTooltip')}>
-                  <PixelIcon name="phone" size={12} /> {t('catalog.cameraBadge')}
-                </span>
-              )}
-            </h3>
-            <p className="card-author">
-              {entry.creatorHandle && !isPlatformAuthor(entry.submittedBy) ? (
-                <>
-                  {t('player.byAuthorPrefix')}
-                  <a className="card-author-link" href={creatorPath(entry.creatorHandle)}>
-                    {entry.submittedBy}
-                  </a>
-                </>
-              ) : (
-                t('player.byAuthor', {
-                  author: isPlatformAuthor(entry.submittedBy) ? t('catalog.platformAuthor') : entry.submittedBy,
-                })
-              )}
-            </p>
-            {/*
-             * Contributor credit — the growth loop, and the reason proposing is worth doing.
-             * Its own line rather than appended to the byline: the owner is who made the
-             * game, and a contributor sharing that sentence would blur authorship the
-             * ownership rules are careful to keep clear.
-             */}
-            {entry.contributorHandles && entry.contributorHandles.length > 0 ? (
-              <p className="card-contributors">
-                {t('catalog.withContributions')}{' '}
-                {entry.contributorHandles.map((handle, index) => (
-                  <span key={handle}>
-                    {index > 0 ? ', ' : null}
-                    <a className="card-author-link" href={creatorPath(handle)}>
-                      @{handle}
-                    </a>
-                  </span>
-                ))}
-              </p>
-            ) : null}
-          </div>
-          <div className="card-actions">
-            <button className="primary-btn" onClick={() => onPlayGame(entry)}>
-              <PixelIcon name="play" size={13} /> {t('catalog.play')}
-            </button>
-            {entry.multiplayer && (
-              <button className="secondary-btn party-btn" onClick={() => onPlayTogether(entry)}>
-                <PixelIcon name="phone" size={13} /> {t('party.playTogether')}
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-    </article>
-  );
-}
 
 function payloadToSignals(payload: CatalogSortPayload): CatalogSortSignals {
   return {
